Fixing up some things that will eventually stop working in Swift 4 because of changes to NSNumber bridging. Probably should have done it this way all along.

+++ Artmospherez/Forecast.swift
@@ -16,11 +16,11 @@ struct Forecast {
     var minTemp : Int
     var weatherImage: WeatherImage?
 
-    init(dayValue: String, weatherID: WeatherKind, humidityValue: Int, maxTempValue: Int, minTempValue: Int) {
+    init(dayValue: String, weatherID: WeatherKind, humidityValue: Double, maxTempValue: Double, minTempValue: Double) {
         day = dayValue
         kind = weatherID
-        humidity = humidityValue
-        maxTemp = maxTempValue
-        minTemp = minTempValue
+        humidity = Int(humidityValue)
+        maxTemp = Int(maxTempValue)
+        minTemp = Int(minTempValue)
     }
 }
